requirements.txt
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dnspython==2.4.1
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
